Read Folders terminated by null-character instead
This is a tricky one, but if you have special
characters inside your loop/string/file then
basename/bash will not be able to process them properly.
Instead we must terminate them by a null characters using the IFS
variable (Input-Field-Seperator). Now we are sure that the string is complete.